extract aggregation test table setup into separate modules/functions
