+++ custom/icds_reports/tests/__init__.py
@@ -2,7 +2,6 @@ import os
 from datetime import datetime
 
 import mock
-import postgres_copy
 import sqlalchemy
 import csv
 
@@ -14,9 +13,8 @@ from corehq.apps.domain.models import Domain
 from corehq.apps.domain.shortcuts import create_domain
 from corehq.apps.locations.models import SQLLocation, LocationType
 from corehq.apps.userreports.models import StaticDataSourceConfiguration
-from corehq.apps.userreports.util import get_indicator_adapter, get_table_name
+from corehq.apps.userreports.util import get_indicator_adapter
 from corehq.sql_db.connections import connection_manager, ICDS_UCR_ENGINE_ID
-from custom.icds_reports.const import DISTRIBUTED_TABLES, REFERENCE_TABLES
 
 from custom.icds_reports.tasks import (
     move_ucr_data_into_aggregation_tables,
@@ -24,40 +22,9 @@ from custom.icds_reports.tasks import (
     _aggregate_child_health_pnc_forms,
     _aggregate_bp_forms,
     _aggregate_gm_forms)
-from .agg_setup import setup_location_hierarchy
+from .agg_setup import setup_location_hierarchy, setup_table_fixtures
 
-from custom.icds_reports.utils.migrations import create_citus_reference_table, create_citus_distributed_table
 
-FILE_NAME_TO_TABLE_MAPPING = {
-    'awc_mgmt': get_table_name('icds-cas', 'static-awc_mgt_forms'),
-    "ccs_cases": get_table_name('icds-cas', 'static-ccs_record_cases'),
-    'child_cases': get_table_name('icds-cas', 'static-child_health_cases'),
-    'daily_feeding': get_table_name('icds-cas', 'static-daily_feeding_forms'),
-    'household_cases': get_table_name('icds-cas', 'static-household_cases'),
-    'infrastructure': get_table_name('icds-cas', 'static-infrastructure_form'),
-    'infrastructure_v2': get_table_name('icds-cas', 'static-infrastructure_form_v2'),
-    'person_cases': get_table_name('icds-cas', 'static-person_cases_v3'),
-    'usage': get_table_name('icds-cas', 'static-usage_forms'),
-    'vhnd': get_table_name('icds-cas', 'static-vhnd_form'),
-    'complementary_feeding': get_table_name('icds-cas', 'static-complementary_feeding_forms'),
-    'aww_user': get_table_name('icds-cas', 'static-commcare_user_cases'),
-    'child_tasks': get_table_name('icds-cas', 'static-child_tasks_cases'),
-    'pregnant_tasks': get_table_name('icds-cas', 'static-pregnant-tasks_cases'),
-    'thr_form': get_table_name('icds-cas', 'static-dashboard_thr_forms'),
-    'gm_form': get_table_name('icds-cas', 'static-dashboard_growth_monitoring_forms'),
-    'pnc_forms': get_table_name('icds-cas', 'static-postnatal_care_forms'),
-    'dashboard_daily_feeding': get_table_name('icds-cas', 'dashboard_child_health_daily_feeding_forms'),
-    'ls_awc_mgt': get_table_name('icds-cas', 'static-awc_mgt_forms'),
-    'ls_home_vists': get_table_name('icds-cas', 'static-ls_home_visit_forms_filled'),
-    'ls_vhnd': get_table_name('icds-cas', 'static-ls_vhnd_form'),
-    'cbe_form': get_table_name('icds-cas', 'static-cbe_form'),
-    'birth_preparedness': get_table_name('icds-cas', 'static-dashboard_birth_preparedness_forms'),
-    'delivery_form': get_table_name('icds-cas', 'static-dashboard_delivery_forms'),
-    'thr_form_v2': get_table_name('icds-cas', 'static-thr_forms_v2'),
-    'awc_location': 'awc_location',
-    'awc_location_local': 'awc_location_local',
-    'agg_awc': 'agg_awc',
-}
 
 OUTPUT_PATH = os.path.join(os.path.dirname(__file__), 'outputs')
 
@@ -79,33 +46,7 @@ def setUpModule():
     setup_location_hierarchy(domain.name)
 
     with override_settings(SERVER_ENVIRONMENT='icds'):
-        configs = StaticDataSourceConfiguration.by_domain('icds-cas')
-        adapters = [get_indicator_adapter(config) for config in configs]
-
-        for adapter in adapters:
-            try:
-                adapter.drop_table()
-            except Exception:
-                pass
-            adapter.build_table()
-
-        engine = connection_manager.get_engine(ICDS_UCR_ENGINE_ID)
-        metadata = sqlalchemy.MetaData(bind=engine)
-        metadata.reflect(bind=engine, extend_existing=True)
-        path = os.path.join(os.path.dirname(__file__), 'fixtures')
-        for file_name in os.listdir(path):
-            with open(os.path.join(path, file_name), encoding='utf-8') as f:
-                table_name = FILE_NAME_TO_TABLE_MAPPING[file_name[:-4]]
-                table = metadata.tables[table_name]
-                if not table_name.startswith('icds_dashboard_'):
-                    columns = [
-                        '"{}"'.format(c.strip())  # quote to preserve case
-                        for c in f.readline().split(',')
-                    ]
-                    postgres_copy.copy_from(f, table, engine, format='csv', null='', columns=columns)
-
-        _distribute_tables_for_citus(engine)
-
+        setup_table_fixtures()
         for state_id in ('st1', 'st2'):
             _aggregate_child_health_pnc_forms(state_id, datetime(2017, 3, 31))
             _aggregate_gm_forms(state_id, datetime(2017, 3, 31))
@@ -121,35 +62,6 @@ def setUpModule():
             raise
         finally:
             _call_center_domain_mock.stop()
-
-
-def _distribute_tables_for_citus(engine):
-    if not getattr(settings, 'ICDS_USE_CITUS', False):
-        return
-
-    for table, col in DISTRIBUTED_TABLES:
-        with engine.begin() as conn:
-
-            # TODO: remove this after citus migration
-            res = conn.execute(
-                """
-                SELECT c.relname AS child
-                FROM
-                    pg_inherits JOIN pg_class AS c ON (inhrelid=c.oid)
-                    JOIN pg_class as p ON (inhparent=p.oid)
-                    where p.relname = %s;
-                """,
-                table
-            )
-            for child in [row.child for row in res]:
-                # only need this because of reusedb if testing on master and this branch
-                conn.execute('drop table if exists "{}"'.format(child))
-
-            create_citus_distributed_table(conn, table, col)
-
-    for table in REFERENCE_TABLES:
-        with engine.begin() as conn:
-            create_citus_reference_table(conn, table)
 
 
 def tearDownModule():
